Switch to repository for cookie (in case it's needed)
Please ignore this commit if we decide not to do it this way
or squash the previous one if we do like it.
